+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - The GUI is not locked anymore while a popup is shown.
 - The tag preferences have been updated.
 - CodeMirror by default sets the cursor to the beginning or end of a whole line (with line wrapping). You can now change that behaviour, so that the `Home` and `End` buttons bring you to the beginning and end of the _visible_ lines, not the logical lines.
+- Zettlr now counts completed pomodoros.
 
 ### Fixes